Rearrange headers
