+++ include/Makefile.am
@@ -1,3 +1,3 @@
 polycapincludedir = ${includedir}/polycap
 
-polycapinclude_HEADERS = polycap.h polycap-error.h polycap-profile.h polycap-description.h
+polycapinclude_HEADERS = polycap.h polycap-error.h polycap-profile.h polycap-description.h polycap-source.h polycap-rng.h polycap-transmission-efficiencies.h polycap-photon.h
